requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.9
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
dbus-python==1.2.18
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
flake8==7.3.0
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.5.1
future==1.0.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib-metadata==4.6.4
isort==4.3.21
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
Mako==1.1.3
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
meson==1.5.2
more-itertools==8.10.0
mpmath==1.3.0
narwhals==1.47.1
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.18.0
onnxruntime==1.22.0
onnxslim==0.1.59
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
portalocker==2.10.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyds @ file:///opt/nvidia/deepstream/deepstream-7.1/pyds-1.2.0-cp310-cp310-linux_x86_64.whl
pyflakes==3.4.0
Pygments==2.11.2
PyGObject==3.42.1
pyparsing==3.2.3
pyservicemaker @ file:///opt/nvidia/deepstream/deepstream-7.1/service-maker/python/pyservicemaker-0.0.1-py3-none-linux_x86_64.whl
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dbusmock==0.27.5
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.3
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.1
streamlit==1.47.0
streamlit-autorefresh==1.0.1
sympy==1.14.0
tb-nightly==2.20.0a20250709
tenacity==9.1.2
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
torch==2.7.1
-e git+https://github.com/KaiyangZhou/deep-person-reid.git@566a56a2cb255f59ba75aa817032621784df546a#egg=torchreid
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.163
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
Werkzeug==3.1.3
yacs==0.1.8
yapf==0.43.0
zipp==1.0.0
